src/core/jupiter_trader.py
import requests

class JupiterTrader:
    def __init__(self, api_url):
        self.api_url = api_url

    def execute_trade(self, token_address, amount):
        # Placeholder for executing a trade through the Jupiter API
        payload = {
            'token_address': token_address,
            'amount': amount
        }
        response = requests.post(f'{self.api_url}/trade', json=payload)
        if response.status_code == 200:
            print(f'Trade executed: {response.json()}')
            return response.json()  # Return the response
        else:
            print(f'Error executing trade: {response.status_code} - {response.text}')
            return None  # Return None on error

    def get_trade_history(self):
        # Placeholder for getting trade history from the Jupiter API
        return [{"token_address": "token_address", "amount": 100, "timestamp": "2023-01-01T00:00:00"}]

